mem-ruby: Fix type casting in makeNextStrideAddress.
This causes the RubyPrefetcher failed to calculate
correct prefetched address.

+++ src/mem/ruby/common/Address.cc
@@ -53,7 +53,8 @@ makeLineAddress(Addr addr)
 Addr
 makeNextStrideAddress(Addr addr, int stride)
 {
-    return makeLineAddress(addr) + RubySystem::getBlockSizeBytes() * stride;
+    return makeLineAddress(addr) +
+        static_cast<int>(RubySystem::getBlockSizeBytes()) * stride;
 }
 
 std::string
